[OP] Move reshape and flatten to nnvm (#4101)
* Move reshape and flatten to NNVM

* Add test for Flatten
